lumiere/urls.py
from django.contrib import admin
from django.urls import include, path

from drf_spectacular.views import (
    SpectacularAPIView,
    SpectacularRedocView,
    SpectacularSwaggerView,
)
from rest_framework.routers import DefaultRouter
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView

from core.views import (
    ArtistasViewSet,
    CriticaViewSet,
    FilmesViewSet,
    GeneroViewSet,
    PlataformaViewSet,
)

router = DefaultRouter()
router.register(r'filmes', FilmesViewSet)
router.register(r'genero', GeneroViewSet)
router.register(r'artistas', ArtistasViewSet)
router.register(r'plataforma', PlataformaViewSet)
router.register(r'critica', CriticaViewSet)

urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/', include(router.urls)),
    path("token/", TokenObtainPairView.as_view(), name="token_obtain_pair"),
    path("token/refresh/", TokenRefreshView.as_view(), name="token_refresh"),
    path("api/schema/", SpectacularAPIView.as_view(), name="schema"),
    path(
        "api/swagger/",
        SpectacularSwaggerView.as_view(url_name="schema"),
        name="swagger-ui",
    ),
    path(
        "api/redoc/",
        SpectacularRedocView.as_view(url_name="schema"),
        name="redoc",
    ),
]