pyproject.toml
[tool.poetry]
name = "lumiere"
version = "0.1.0"
description = ""
authors = ["marianaabento <mariibentoo04@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1"
djangorestframework = "^3.13.1"
gunicorn = "^20.1.0"
whitenoise = "^6.2.0"
psycopg2-binary = "^2.9.4"
djangorestframework-simplejwt = "^5.2.1"
drf-spectacular = "^0.24.2"
isort = "^5.10.1"
django-decouple = "^2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
